[TWRP][0.0.1.0] Вынос функций в отдельный модуль
