+++ conf.yml
@@ -1,2 +1,2 @@
-version: 0.0.0.2
-url:
+version: 0.0.1.0
+url: https://... # URL to audio stream
